add Dockerfile, protect against Spam, ...
The environment variable CHAT_ID can now be used to limit the bot to a single chat. If you don't set the CHAT_ID, the client can set any chatId to send a message and therefore *abuse* your Bot.
With the Dockerfile, an extrem easy deployment is now possible.
Improved dokumentation.
some cleanup